Initial support or automatic reload on file change based on QFileSystemWatcher.
This actually doesn't work but yet the ideal example has been committed but disabled.  Complications are multiple file change events coming in for what is logically one change and change events coming in before the change is finished and before the file is actually a valid HDF4/HDF5/NetCDF file.  I know there is a solution as okular has this feature for PDF files which I rely on for Latex writing.